96252: Extract form-specific code from SharedModule
