+++ src/app/shared/form/chips/models/chips-item.model.spec.ts
@@ -1,5 +1,5 @@
 import { ChipsItem, ChipsItemIcon } from './chips-item.model';
-import { FormFieldMetadataValueObject } from '../../form/builder/models/form-field-metadata-value.model';
+import { FormFieldMetadataValueObject } from '../../builder/models/form-field-metadata-value.model';
 
 describe('ChipsItem model test suite', () => {
   let item: ChipsItem;
