+++ src/app/shared/form/chips/models/chips.model.ts
@@ -1,11 +1,11 @@
 import { findIndex, isEqual, isObject } from 'lodash';
 import { BehaviorSubject } from 'rxjs';
 import { ChipsItem, ChipsItemIcon } from './chips-item.model';
-import { hasValue, isNotEmpty } from '../../empty.util';
-import { MetadataIconConfig } from '../../../../config/submission-config.interface';
-import { FormFieldMetadataValueObject } from '../../form/builder/models/form-field-metadata-value.model';
-import { VocabularyEntry } from '../../../core/submission/vocabularies/models/vocabulary-entry.model';
-import { PLACEHOLDER_PARENT_METADATA } from '../../form/builder/ds-dynamic-form-ui/ds-dynamic-form-constants';
+import { hasValue, isNotEmpty } from '../../../empty.util';
+import { MetadataIconConfig } from '../../../../../config/submission-config.interface';
+import { FormFieldMetadataValueObject } from '../../builder/models/form-field-metadata-value.model';
+import { VocabularyEntry } from '../../../../core/submission/vocabularies/models/vocabulary-entry.model';
+import { PLACEHOLDER_PARENT_METADATA } from '../../builder/ds-dynamic-form-ui/ds-dynamic-form-constants';
 
 export class Chips {
   chipsItems: BehaviorSubject<ChipsItem[]>;
