+++ src/app/shared/form/number-picker/number-picker.component.scss
