+++ src/app/shared/form/vocabulary-treeview/vocabulary-tree-flattener.ts
