+++ src/app/shared/form/vocabulary-treeview/vocabulary-treeview.component.scss
