+++ src/app/submit-page/submit-page.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
 import { SubmitPageRoutingModule } from './submit-page-routing.module';
 import { SubmissionModule } from '../submission/submission.module';
+import { FormModule } from '../shared/form/form.module';
 
 @NgModule({
   imports: [
@@ -10,6 +11,7 @@ import { SubmissionModule } from '../submission/submission.module';
     CommonModule,
     SharedModule,
     SubmissionModule,
+    FormModule,
   ],
 })
 /**
